+++ src/lib/Select.svelte
@@ -5,7 +5,6 @@
     import _Item from './Item.svelte';
     import _Selection from './Selection.svelte';
     import _MultiSelection from './MultiSelection.svelte';
-    import _VirtualList from './VirtualList.svelte';
     import _ClearIcon from './ClearIcon.svelte';
     import debounce from './debounce';
 
@@ -71,7 +70,6 @@
     export let isWaiting = false;
     export let listPlacement = 'auto';
     export let listOpen = false;
-    export let isVirtualList = false;
     export let loadOptionsInterval = 300;
     export let noOptionsMessage = 'No options';
     export let hideEmptyState = false;
@@ -91,7 +89,7 @@
     export let List = _List;
     export let Selection = _Selection;
     export let MultiSelection = _MultiSelection;
-    export let VirtualList = _VirtualList;
+    export let VirtualList = null;
 
     function filterMethod(args) {
         if (args.loadOptions && args.filterText.length > 0) return;
@@ -570,7 +568,6 @@
         optionIdentifier,
         noOptionsMessage,
         hideEmptyState,
-        isVirtualList,
         VirtualList,
         value,
         isMulti,